[dfs] Fixing bugs & refactoring
Dangling reference in wrappers were removed
Introduce DFS Test interface
bb_id trait type added
